python-server/requirements.txt
Flask
Werkzeug
Flask-RESTx
openai
PyPDF2
python-dotenv
python-docx